refactor: models package to entities
Tested-by: Nara <98963440+vianneynara@users.noreply.github.com>

Took 4 minutes
