+++ src/main/java/dev/realtards/kuenyawz/configurations/AccountUserDetails.java
@@ -1,6 +1,6 @@
 package dev.realtards.kuenyawz.configurations;
 
-import dev.realtards.kuenyawz.models.Account;
+import dev.realtards.kuenyawz.entities.Account;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
