+++ src/main/java/dev/realtards/kuenyawz/entities/Auditables.java
@@ -1,4 +1,4 @@
-package dev.realtards.kuenyawz.models;
+package dev.realtards.kuenyawz.entities;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
