+++ src/main/java/dev/realtards/kuenyawz/entities/ProductVariant.java
@@ -1,4 +1,4 @@
-package dev.realtards.kuenyawz.models;
+package dev.realtards.kuenyawz.entities;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
